refactor: Rename old IntegrationPerformanceTestRunner to Legacy and replace with V2 version
