+++ src/test/java/com/synopsys/integration/alert/performance/JiraCloudPerformanceTest.java
@@ -34,7 +34,7 @@ import com.synopsys.integration.alert.descriptor.api.JiraCloudChannelKey;
 import com.synopsys.integration.alert.performance.utility.AlertRequestUtility;
 import com.synopsys.integration.alert.performance.utility.BlackDuckProviderService;
 import com.synopsys.integration.alert.performance.utility.ConfigurationManager;
-import com.synopsys.integration.alert.performance.utility.IntegrationPerformanceTestRunner;
+import com.synopsys.integration.alert.performance.utility.IntegrationPerformanceTestRunnerLegacy;
 import com.synopsys.integration.alert.test.common.TestProperties;
 import com.synopsys.integration.alert.test.common.TestPropertyKey;
 import com.synopsys.integration.alert.test.common.TestTags;
@@ -55,8 +55,8 @@ public class JiraCloudPerformanceTest {
     @Autowired
     private JiraCloudCustomFunctionAction jiraCloudCustomFunctionAction;
 
-    private final Gson gson = IntegrationPerformanceTestRunner.createGson();
-    private final DateTimeFormatter dateTimeFormatter = IntegrationPerformanceTestRunner.createDateTimeFormatter();
+    private final Gson gson = IntegrationPerformanceTestRunnerLegacy.createGson();
+    private final DateTimeFormatter dateTimeFormatter = IntegrationPerformanceTestRunnerLegacy.createDateTimeFormatter();
 
     @Test
     @Disabled
@@ -72,16 +72,20 @@ public class JiraCloudPerformanceTest {
 
         Map<String, FieldValueModel> channelFieldsMap = createChannelFieldsMap(testProperties);
 
-        IntegrationPerformanceTestRunner testRunner = createTestRunner();
+        IntegrationPerformanceTestRunnerLegacy testRunner = createTestRunner();
         testRunner.runTest(globalConfig, channelFieldsMap, JIRA_CLOUD_PERFORMANCE_JOB_NAME);
     }
 
     private FieldModel createGlobalConfig(TestProperties testProperties) {
         Map<String, FieldValueModel> globalConfigFields = Map.of(
-            JiraCloudDescriptor.KEY_JIRA_URL, createFieldValueModel(testProperties.getProperty(TestPropertyKey.TEST_JIRA_CLOUD_URL)),
-            JiraCloudDescriptor.KEY_JIRA_ADMIN_EMAIL_ADDRESS, createFieldValueModel(testProperties.getProperty(TestPropertyKey.TEST_JIRA_CLOUD_USER_EMAIL)),
-            JiraCloudDescriptor.KEY_JIRA_ADMIN_API_TOKEN, createFieldValueModel(testProperties.getProperty(TestPropertyKey.TEST_JIRA_CLOUD_API_TOKEN)),
-            JiraCloudDescriptor.KEY_JIRA_DISABLE_PLUGIN_CHECK, createFieldValueModel(testProperties.getOptionalProperty(TestPropertyKey.TEST_JIRA_CLOUD_DISABLE_PLUGIN_CHECK).orElse("false"))
+            JiraCloudDescriptor.KEY_JIRA_URL,
+            createFieldValueModel(testProperties.getProperty(TestPropertyKey.TEST_JIRA_CLOUD_URL)),
+            JiraCloudDescriptor.KEY_JIRA_ADMIN_EMAIL_ADDRESS,
+            createFieldValueModel(testProperties.getProperty(TestPropertyKey.TEST_JIRA_CLOUD_USER_EMAIL)),
+            JiraCloudDescriptor.KEY_JIRA_ADMIN_API_TOKEN,
+            createFieldValueModel(testProperties.getProperty(TestPropertyKey.TEST_JIRA_CLOUD_API_TOKEN)),
+            JiraCloudDescriptor.KEY_JIRA_DISABLE_PLUGIN_CHECK,
+            createFieldValueModel(testProperties.getOptionalProperty(TestPropertyKey.TEST_JIRA_CLOUD_DISABLE_PLUGIN_CHECK).orElse("false"))
         );
         return new FieldModel(CHANNEL_KEY.getUniversalKey(), ConfigContextEnum.GLOBAL.name(), globalConfigFields);
     }
@@ -94,12 +98,21 @@ public class JiraCloudPerformanceTest {
         channelFieldsMap.put(ChannelDescriptor.KEY_FREQUENCY, createFieldValueModel(FrequencyType.REAL_TIME.name()));
         channelFieldsMap.put(ChannelDescriptor.KEY_PROVIDER_TYPE, createFieldValueModel(PROVIDER_KEY.getUniversalKey()));
 
-        channelFieldsMap.put(JiraCloudDescriptor.KEY_ADD_COMMENTS, createFieldValueModel(testProperties.getOptionalProperty(TestPropertyKey.TEST_JIRA_CLOUD_ADD_COMMENTS).orElse("true")));
-        channelFieldsMap.put(JiraCloudDescriptor.KEY_ISSUE_CREATOR, createFieldValueModel(testProperties.getOptionalProperty(TestPropertyKey.TEST_JIRA_CLOUD_ISSUE_CREATOR).orElse("")));
+        channelFieldsMap
+            .put(JiraCloudDescriptor.KEY_ADD_COMMENTS, createFieldValueModel(testProperties.getOptionalProperty(TestPropertyKey.TEST_JIRA_CLOUD_ADD_COMMENTS).orElse("true")));
+        channelFieldsMap
+            .put(JiraCloudDescriptor.KEY_ISSUE_CREATOR, createFieldValueModel(testProperties.getOptionalProperty(TestPropertyKey.TEST_JIRA_CLOUD_ISSUE_CREATOR).orElse("")));
         channelFieldsMap.put(JiraCloudDescriptor.KEY_JIRA_PROJECT_NAME, createFieldValueModel(testProperties.getProperty(TestPropertyKey.TEST_JIRA_CLOUD_PROJECT_NAME)));
-        channelFieldsMap.put(JiraCloudDescriptor.KEY_ISSUE_TYPE, createFieldValueModel(testProperties.getOptionalProperty(TestPropertyKey.TEST_JIRA_CLOUD_ISSUE_TYPE).orElse("Task")));
-        channelFieldsMap.put(JiraCloudDescriptor.KEY_RESOLVE_WORKFLOW_TRANSITION, createFieldValueModel(testProperties.getOptionalProperty(TestPropertyKey.TEST_JIRA_CLOUD_RESOLVE_TRANSITION).orElse("")));
-        channelFieldsMap.put(JiraCloudDescriptor.KEY_OPEN_WORKFLOW_TRANSITION, createFieldValueModel(testProperties.getOptionalProperty(TestPropertyKey.TEST_JIRA_CLOUD_REOPEN_TRANSITION).orElse("")));
+        channelFieldsMap
+            .put(JiraCloudDescriptor.KEY_ISSUE_TYPE, createFieldValueModel(testProperties.getOptionalProperty(TestPropertyKey.TEST_JIRA_CLOUD_ISSUE_TYPE).orElse("Task")));
+        channelFieldsMap.put(
+            JiraCloudDescriptor.KEY_RESOLVE_WORKFLOW_TRANSITION,
+            createFieldValueModel(testProperties.getOptionalProperty(TestPropertyKey.TEST_JIRA_CLOUD_RESOLVE_TRANSITION).orElse(""))
+        );
+        channelFieldsMap.put(
+            JiraCloudDescriptor.KEY_OPEN_WORKFLOW_TRANSITION,
+            createFieldValueModel(testProperties.getOptionalProperty(TestPropertyKey.TEST_JIRA_CLOUD_REOPEN_TRANSITION).orElse(""))
+        );
 
         return channelFieldsMap;
     }
@@ -108,11 +121,16 @@ public class JiraCloudPerformanceTest {
         return new FieldValueModel(List.of(value), true);
     }
 
-    private IntegrationPerformanceTestRunner createTestRunner() {
-        AlertRequestUtility alertRequestUtility = IntegrationPerformanceTestRunner.createAlertRequestUtility(webApplicationContext);
+    private IntegrationPerformanceTestRunnerLegacy createTestRunner() {
+        AlertRequestUtility alertRequestUtility = IntegrationPerformanceTestRunnerLegacy.createAlertRequestUtility(webApplicationContext);
         BlackDuckProviderService blackDuckProviderService = new BlackDuckProviderService(alertRequestUtility, gson);
-        ConfigurationManager configurationManager = new ConfigurationManager(gson, alertRequestUtility, blackDuckProviderService.getBlackDuckProviderKey(), CHANNEL_KEY.getUniversalKey());
-        return new IntegrationPerformanceTestRunner(gson, dateTimeFormatter, alertRequestUtility, blackDuckProviderService, configurationManager);
+        ConfigurationManager configurationManager = new ConfigurationManager(
+            gson,
+            alertRequestUtility,
+            blackDuckProviderService.getBlackDuckProviderKey(),
+            CHANNEL_KEY.getUniversalKey()
+        );
+        return new IntegrationPerformanceTestRunnerLegacy(gson, dateTimeFormatter, alertRequestUtility, blackDuckProviderService, configurationManager);
     }
 
 }
