+++ src/test/java/com/synopsys/integration/alert/performance/LargeNotificationTest.java
@@ -33,7 +33,7 @@ import com.synopsys.integration.alert.performance.model.PerformanceExecutionStat
 import com.synopsys.integration.alert.performance.utility.AlertRequestUtility;
 import com.synopsys.integration.alert.performance.utility.BlackDuckProviderService;
 import com.synopsys.integration.alert.performance.utility.ConfigurationManagerV2;
-import com.synopsys.integration.alert.performance.utility.IntegrationPerformanceTestRunnerV2;
+import com.synopsys.integration.alert.performance.utility.IntegrationPerformanceTestRunner;
 import com.synopsys.integration.alert.performance.utility.jira.server.JiraServerPerformanceUtility;
 import com.synopsys.integration.alert.test.common.TestProperties;
 import com.synopsys.integration.alert.test.common.TestPropertyKey;
@@ -60,19 +60,19 @@ class LargeNotificationTest {
     private WebApplicationContext webApplicationContext;
 
     private final Logger logger = LoggerFactory.getLogger(getClass());
-    private final Gson gson = IntegrationPerformanceTestRunnerV2.createGson();
-    private final DateTimeFormatter dateTimeFormatter = IntegrationPerformanceTestRunnerV2.createDateTimeFormatter();
+    private final Gson gson = IntegrationPerformanceTestRunner.createGson();
+    private final DateTimeFormatter dateTimeFormatter = IntegrationPerformanceTestRunner.createDateTimeFormatter();
 
     private BlackDuckProviderService blackDuckProviderService;
     private JiraServerPerformanceUtility jiraServerPerformanceUtility;
-    private IntegrationPerformanceTestRunnerV2 testRunner;
+    private IntegrationPerformanceTestRunner testRunner;
 
     private TestProperties testProperties = new TestProperties();
     private int numberOfProjectsToCreate;
 
     @BeforeEach
     public void init() {
-        AlertRequestUtility alertRequestUtility = IntegrationPerformanceTestRunnerV2.createAlertRequestUtility(webApplicationContext);
+        AlertRequestUtility alertRequestUtility = IntegrationPerformanceTestRunner.createAlertRequestUtility(webApplicationContext);
         blackDuckProviderService = new BlackDuckProviderService(alertRequestUtility, gson);
         ConfigurationManagerV2 configurationManager = new ConfigurationManagerV2(
             gson,
@@ -81,7 +81,7 @@ class LargeNotificationTest {
             CHANNEL_KEY.getUniversalKey()
         );
         jiraServerPerformanceUtility = new JiraServerPerformanceUtility(alertRequestUtility, configurationManager);
-        testRunner = new IntegrationPerformanceTestRunnerV2(
+        testRunner = new IntegrationPerformanceTestRunner(
             gson,
             dateTimeFormatter,
             alertRequestUtility,
