+++ src/test/java/com/synopsys/integration/alert/performance/MsTeamsPerformanceTest.java
@@ -29,7 +29,7 @@ import com.synopsys.integration.alert.descriptor.api.model.ChannelKeys;
 import com.synopsys.integration.alert.performance.utility.AlertRequestUtility;
 import com.synopsys.integration.alert.performance.utility.BlackDuckProviderService;
 import com.synopsys.integration.alert.performance.utility.ConfigurationManager;
-import com.synopsys.integration.alert.performance.utility.IntegrationPerformanceTestRunner;
+import com.synopsys.integration.alert.performance.utility.IntegrationPerformanceTestRunnerLegacy;
 import com.synopsys.integration.alert.test.common.TestProperties;
 import com.synopsys.integration.alert.test.common.TestPropertyKey;
 import com.synopsys.integration.alert.test.common.TestTags;
@@ -46,8 +46,8 @@ public class MsTeamsPerformanceTest {
     @Autowired
     private WebApplicationContext webApplicationContext;
 
-    private final Gson gson = IntegrationPerformanceTestRunner.createGson();
-    private final DateTimeFormatter dateTimeFormatter = IntegrationPerformanceTestRunner.createDateTimeFormatter();
+    private final Gson gson = IntegrationPerformanceTestRunnerLegacy.createGson();
+    private final DateTimeFormatter dateTimeFormatter = IntegrationPerformanceTestRunnerLegacy.createDateTimeFormatter();
 
     private static String MSTEAMS_CHANNEL_WEBHOOK;
 
@@ -61,7 +61,7 @@ public class MsTeamsPerformanceTest {
     @Ignore // performance test
     @Disabled
     void testMsTeamsJob() throws Exception {
-        AlertRequestUtility alertRequestUtility = IntegrationPerformanceTestRunner.createAlertRequestUtility(webApplicationContext);
+        AlertRequestUtility alertRequestUtility = IntegrationPerformanceTestRunnerLegacy.createAlertRequestUtility(webApplicationContext);
         BlackDuckProviderService blackDuckProviderService = new BlackDuckProviderService(alertRequestUtility, gson);
         ConfigurationManager configurationManager = new ConfigurationManager(
             gson,
@@ -69,7 +69,7 @@ public class MsTeamsPerformanceTest {
             blackDuckProviderService.getBlackDuckProviderKey(),
             ChannelKeys.MS_TEAMS.getUniversalKey()
         );
-        IntegrationPerformanceTestRunner integrationPerformanceTestRunner = new IntegrationPerformanceTestRunner(
+        IntegrationPerformanceTestRunnerLegacy integrationPerformanceTestRunner = new IntegrationPerformanceTestRunnerLegacy(
             gson,
             dateTimeFormatter,
             alertRequestUtility,
